refactor: move and rename stories
